+++ Apps/Examples/Examples/All Examples/FitCameraToGeometryExample.swift
@@ -38,8 +38,8 @@ public class FitCameraToGeometryExample: UIViewController, ExampleProtocol {
         source.data = .feature(polygonFeature)
 
         var polygonLayer = FillLayer(id: "triangle-style")
-        polygonLayer.paint?.fillOpacity = .constant(0.5)
-        polygonLayer.paint?.fillColor = .constant(ColorRepresentable(color: .gray))
+        polygonLayer.fillOpacity = .constant(0.5)
+        polygonLayer.fillColor = .constant(ColorRepresentable(color: .gray))
         polygonLayer.source = sourceIdentifier
 
         do {