+++ Apps/Examples/Examples/All Examples/GeoJSONSourceExample.swift
@@ -65,9 +65,9 @@ public class GeoJSONSourceExample: UIViewController, ExampleProtocol {
             "Point"
         }
         circleLayer.source = geoJSONDataSourceIdentifier
-        circleLayer.paint?.circleColor = .constant(ColorRepresentable(color: UIColor.yellow))
-        circleLayer.paint?.circleOpacity = .constant(0.6)
-        circleLayer.paint?.circleRadius = .constant(8.0)
+        circleLayer.circleColor = .constant(ColorRepresentable(color: UIColor.yellow))
+        circleLayer.circleOpacity = .constant(0.6)
+        circleLayer.circleRadius = .constant(8.0)
         // Follow the same steps to create a line layer
         var lineLayer = LineLayer(id: "line-layer")
         lineLayer.filter = Exp(.eq) {
@@ -75,8 +75,8 @@ public class GeoJSONSourceExample: UIViewController, ExampleProtocol {
             "LineString"
         }
         lineLayer.source = geoJSONDataSourceIdentifier
-        lineLayer.paint?.lineColor = .constant(ColorRepresentable(color: UIColor.red))
-        lineLayer.paint?.lineWidth = .constant(1.4)
+        lineLayer.lineColor = .constant(ColorRepresentable(color: UIColor.red))
+        lineLayer.lineWidth = .constant(1.4)
         // Follow the same steps to create a polygon (fill) layer
         var polygonLayer = FillLayer(id: "fill-layer")
         polygonLayer.filter = Exp(.eq) {
@@ -84,9 +84,9 @@ public class GeoJSONSourceExample: UIViewController, ExampleProtocol {
             "Polygon"
         }
         polygonLayer.source = geoJSONDataSourceIdentifier
-        polygonLayer.paint?.fillColor = .constant(ColorRepresentable(color: UIColor.green))
-        polygonLayer.paint?.fillOpacity = .constant(0.3)
-        polygonLayer.paint?.fillOutlineColor = .constant(ColorRepresentable(color: UIColor.purple))
+        polygonLayer.fillColor = .constant(ColorRepresentable(color: UIColor.green))
+        polygonLayer.fillOpacity = .constant(0.3)
+        polygonLayer.fillOutlineColor = .constant(ColorRepresentable(color: UIColor.purple))
         // Add the source and style layers to the map style.
         try! mapView.mapboxMap.style.addSource(geoJSONSource, id: geoJSONDataSourceIdentifier)
         try! mapView.mapboxMap.style.addLayer(circleLayer)