+++ Apps/Examples/Examples/All Examples/TerrainExample.swift
@@ -41,9 +41,9 @@ public class TerrainExample: UIViewController, ExampleProtocol {
         try! mapView.mapboxMap.style.setTerrain(terrain)
 
         var skyLayer = SkyLayer(id: "sky-layer")
-        skyLayer.paint?.skyType = .constant(.atmosphere)
-        skyLayer.paint?.skyAtmosphereSun = .constant([0.0, 0.0])
-        skyLayer.paint?.skyAtmosphereSunIntensity = .constant(15.0)
+        skyLayer.skyType = .constant(.atmosphere)
+        skyLayer.skyAtmosphereSun = .constant([0.0, 0.0])
+        skyLayer.skyAtmosphereSunIntensity = .constant(15.0)
 
         try! mapView.mapboxMap.style.addLayer(skyLayer)
     }